% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{binomdist}
\alias{binomdist}
\title{Creates an Object Representing Binomial Distribution.}
\usage{
binomdist(size = 10, prob = 0.5)
}
\arguments{
\item{size}{size parameter, default: 10.}

\item{prob}{probability parameter, default: 0.5.}
}
\value{
Object of class binomdist.
}
\description{
The function creates an object which represents the binomial distribution.
}
\details{
See \code{\link[stats]{Binomial}}.
}
\examples{
B <- binomdist(10, 0.4)
d(B, c(2, 3, 4, NA))
r(B, 5)
}
\seealso{
\code{\link[stats]{Binomial}}
}
