% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BA.R
\name{calc_BA}
\alias{calc_BA}
\title{Calculate basal area of a stand.}
\usage{
calc_BA(SI, BHage, SppId, stocking = 1, Origin = "N",
  PlantedSpp = NA)
}
\arguments{
\item{SI}{Site index value}

\item{BHage}{Breast height age}

\item{SppId}{Id of a species or forest unit}

\item{stocking}{Stocking.}

\item{Origin}{Origin of the stand. Use N for natural, P for plantation.}

\item{PlantedSpp}{Opional information on planted species required for some plantations.}
}
\value{
basal area
}
\description{
Calculates basal area of stand based on provided inputs.
}
