% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cull.R
\name{calc_cull}
\alias{calc_cull}
\title{Calculate cull.}
\usage{
calc_cull(age, SppId)
}
\arguments{
\item{age}{stand total age}

\item{SppId}{Id of a species or forest unit}
}
\value{
cull
}
\description{
Calculate cull.
}
