% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_GMV.R
\name{calc_GMV}
\alias{calc_GMV}
\title{Calculate gross merchantable volume}
\usage{
calc_GMV(GTV, QMD, SppId, Origin = "N", PlantedSpp = NA)
}
\arguments{
\item{GTV}{Gross total volume}

\item{QMD}{Quadratic mean diamater}

\item{SppId}{Id of a species or forest unit}

\item{Origin}{Origin of the stand. Use N for natural, P for plantation.}

\item{PlantedSpp}{Opional information on planted species required for some plantations.}
}
\value{
Gross merchantable volume
}
\description{
Calculates gross total volume [m3/ha] of stand based on provided inputs.
}
