% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_GTV.R
\name{calc_GTV}
\alias{calc_GTV}
\title{Calculate gross total volume}
\usage{
calc_GTV(BA, H, SppId, Origin = "N", PlantedSpp = NA)
}
\arguments{
\item{BA}{Basal area}

\item{H}{Top height}

\item{SppId}{Id of a species or forest unit}

\item{Origin}{Origin of the stand. Use N for natural, P for plantation.}

\item{PlantedSpp}{Opional information on planted species required for some plantations.}
}
\value{
gross total volume
}
\description{
Calculates gross total volume [m3/ha] of stand based on provided inputs.
}
