% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_H.R
\name{calc_H}
\alias{calc_H}
\title{Calculate top height of a stand.}
\usage{
calc_H(SI, age, SppId)
}
\arguments{
\item{SI}{Site index value}

\item{age}{Stand age}

\item{SppId}{Species id.}
}
\value{
Top height.
}
\description{
Calculates top height of a stand based on site index,
age (not BHA), and species.
See technical documentation for MIST for further details on top height
models used for different species
}
