% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_N.R
\name{calc_N}
\alias{calc_N}
\title{Calculate stand density (trees/ha)}
\usage{
calc_N(SI, BHage, BA, SppId, stocking = 1, Origin = "N",
  PlantedSpp = NA)
}
\arguments{
\item{SI}{Site index value}

\item{BHage}{Breast height age}

\item{SppId}{Id of a species or forest unit}

\item{stocking}{Stocking.}

\item{Origin}{Origin of the stand. Use N for natural, P for plantation.}

\item{PlantedSpp}{Opional information on planted species required for some plantations.}
}
\value{
basal area
}
\description{
Calculates density of stand based on provided inputs.
}
