% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NMV.R
\name{calc_NMV}
\alias{calc_NMV}
\title{Calculate net merchantable volume}
\usage{
calc_NMV(cull, GMV)
}
\arguments{
\item{cull}{cull coefficient}

\item{GMV}{Gross merchantable volume}
}
\value{
Gross merchantable volume
}
\description{
Calculates net total volume [m3/ha] of stand based on provided inputs.
}
