% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_QMD.R
\name{calc_QMD}
\alias{calc_QMD}
\title{Calculate quadratic mean diamater}
\usage{
calc_QMD(BA, N)
}
\arguments{
\item{BA}{Basal area}

\item{N}{Stand density}
}
\value{
Quadratic mean diamater
}
\description{
Calculates quadratic mean diamater of stand based on provided inputs.
}
