% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SI.R
\name{calc_SI}
\alias{calc_SI}
\title{Calculate site index of a stand.}
\usage{
calc_SI(H, age, SppId)
}
\arguments{
\item{H}{Top height}

\item{age}{Stand age}

\item{SppId}{Species id.}
}
\value{
Site index
}
\description{
Calculates site index of a stand based on top height,
age (not BHA), and species.
See technical documentation for MIST for further details on site index
models used for different species.
For some species site index models do not exist. In those cases an iterative search algorithm is used
to determine site index value based on top height equation.
}
