% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_VBAR.R
\name{calc_VBAR}
\alias{calc_VBAR}
\title{Calculate volume (GMV) to basal area ratio (VBAR)}
\usage{
calc_VBAR(SI, age, SppId)
}
\arguments{
\item{SI}{site index}

\item{age}{stand total age}

\item{SppId}{Id of a species or forest unit}
}
\value{
VBAR
}
\description{
Calculates VBAR based on SI and age
}
