% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{cauchydist}
\alias{cauchydist}
\title{Creates an Object Representing Cauchy Distribution.}
\usage{
cauchydist(location = 0, scale = 1)
}
\arguments{
\item{location}{location parameter, default: 0.}

\item{scale}{scale parameter, default: 1.}
}
\value{
Object of class cauchydist.
}
\description{
The function creates an object which represents the Cauchy distribution.
}
\details{
See \code{\link[stats]{Cauchy}}.
}
\examples{
C <- cauchydist(0, 1)
d(C, c(2, 3, 4, NA))
r(C, 5)
}
\seealso{
\code{\link[stats]{Cauchy}}
}
