% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{chisqdist}
\alias{chisqdist}
\title{Creates an Object Representing Chi-Squared Distribution}
\usage{
chisqdist(df = 2)
}
\arguments{
\item{df}{degrees of freedom parameter, default: 2.}
}
\value{
Object of class chisqdist.
}
\description{
The function creates an object which represents the chi-squared distribution.
}
\details{
See \code{\link[stats]{Chisquare}}.
}
\examples{
C <- chisqdist(2)
d(C, c(2, 3, 4, NA))
r(C, 5)
}
\seealso{
\code{\link[stats]{Chisquare}}
}
