% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{distribution}
\alias{distribution}
\alias{distribution.comp_fit}
\title{Extract Distribution of Fitted Model}
\usage{
distribution(O)

\method{distribution}{comp_fit}(O)
}
\arguments{
\item{O}{an object for which the extraction of distribution is meaningful.}
}
\value{
Object representing the distribution.
}
\description{
\code{distribution} is a generic function which extracts the distribution with fitted parameters from fitted objects.
}
