% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{expdist}
\alias{expdist}
\title{Creates an Object Representing Exponential Distribution}
\usage{
expdist(rate = 1)
}
\arguments{
\item{rate}{rate parameter, default: 1.}
}
\value{
Object of class expdist.
}
\description{
The function creates an object which represents the exponential distribution.
}
\details{
See \code{\link[stats]{Exponential}}.
}
\examples{
E <- expdist(1)
d(E, c(2, 3, 4, NA))
r(E, 5)
}
\seealso{
\code{\link[stats]{Exponential}}
}
