% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{frechetdist}
\alias{frechetdist}
\title{Creates an Object Representing Frechet Distribution}
\usage{
frechetdist(loc = 0, scale = 1, shape = 1)
}
\arguments{
\item{loc}{location parameter, default: 0.}

\item{scale}{scale parameter, default: 1.}

\item{shape}{shape parameter, default: 1.}
}
\value{
Object of class frechetdist.
}
\description{
The function creates an object which represents the Frechet distribution.
}
\details{
See \code{\link{Frechet}}.
}
\examples{
Fr <- frechetdist(0, 1, 2)
d(Fr, c(2, 3, 4, NA))
r(Fr, 5)
}
\seealso{
\code{\link{Frechet}}
}
