% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{geomdist}
\alias{geomdist}
\title{Creates an Object Representing Geometric Distribution}
\usage{
geomdist(prob = 0.5)
}
\arguments{
\item{prob}{probability parameter, default: 0.5.}
}
\value{
Object of class geomdist.
}
\description{
The function creates an object which represents the geometric distribution.
}
\details{
See \code{\link[stats]{Geometric}}.
}
\examples{
G <- geomdist(0.5)
d(G, c(2, 3, 4, NA))
r(G, 5)
}
\seealso{
\code{\link[stats]{Geometric}}
}
