% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BHA.R
\name{get_BHA}
\alias{get_BHA}
\title{Get breast-height-age for a given species}
\usage{
get_BHA(SppId)
}
\arguments{
\item{SppId}{Id of a species or forest unit}
}
\value{
Years to breast height age
}
\description{
Function to read the breast-height-age (BHA) for given species or forest unit.
note: the table gives both the age2BH value and the two coefficients b0 and b1 that can be used to
calculate the age2bh. The function currently returns just the pre-calculated age2bh. See technical documentation
for MIST for further details.
}
