% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{GPDdist}
\alias{GPDdist}
\title{Creates an Object Representing Generalized Pareto Distribution}
\usage{
GPDdist(loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{loc}{location parameter, default: 0.}

\item{scale}{scale parameter, default: 1.}

\item{shape}{shape parameter, default: 0.}
}
\value{
Object of class GPDdist.
}
\description{
The function creates an object which represents the generalized Pareto distribution.
}
\details{
See \code{\link{GPD}}.
}
\examples{
G <- GPDdist(0, 1, 0)
d(G, c(2, 3, 4, NA))
r(G, 5)
}
\seealso{
\code{\link{GPD}}
}
