% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.contin}
\alias{is.contin}
\title{Reports whether O is a Continuous Distribution Object}
\usage{
is.contin(O)
}
\arguments{
\item{O}{an object to test.}
}
\description{
Reports whether O is a continuous distribution object.
}
