% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.mixture}
\alias{is.mixture}
\title{Reports whether O is a Mixture Distribution Object}
\usage{
is.mixture(O)
}
\arguments{
\item{O}{an object to test.}
}
\description{
Reports whether O is a mixture distribution object.
}
