% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.standard}
\alias{is.standard}
\title{Reports whether O is a Standard Distribution Object}
\usage{
is.standard(O)
}
\arguments{
\item{O}{an object to test.}
}
\description{
Reports whether O is a standard distribution object.
}
