% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{lnormdist}
\alias{lnormdist}
\title{Creates an Object Representing Log Normal Distribution.}
\usage{
lnormdist(meanlog = 0, sdlog = 1)
}
\arguments{
\item{meanlog}{mean parameter, default: 0.}

\item{sdlog}{standard deviation parameter, default: 1.}
}
\value{
Object of class lnormdist.
}
\description{
The function creates an object which represents the log normal distribution.
}
\details{
See \code{\link[stats]{Lognormal}}.
}
\examples{
L <- lnormdist(0, 1)
d(L, c(2, 3, 4, NA))
r(L, 5)
}
\seealso{
\code{\link[stats]{Lognormal}}
}
