% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{normdist}
\alias{normdist}
\title{Creates an Object Representing Normal Distribution}
\usage{
normdist(mean = 0, sd = 1)
}
\arguments{
\item{mean}{mean parameter, default: 0.}

\item{sd}{standard deviation parameter, default: 1.}
}
\value{
Object of class normdist.
}
\description{
The function creates an object which represents the normal distribution.
}
\details{
See \code{\link[stats]{Normal}}.
}
\examples{
N <- normdist(1, 5)
d(N, c(2, 3, 4, NA))
r(N, 5)
}
\seealso{
\code{\link[stats]{Normal}}
}
