% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SI.R
\name{SI_exhaustive}
\alias{SI_exhaustive}
\title{Calculate site index of a stand base on top height model.}
\usage{
SI_exhaustive(H, age, SppId)
}
\arguments{
\item{H}{Top height}

\item{age}{Stand age}

\item{SppId}{Species id.}
}
\value{
Site index
}
\description{
A helper function to use within \code{calc_SI}. Calculates site index of a stand based on top height model,
age (not BHA), and species.
Uses a search algorithm to find the site index value, based on top height equations.
}
