% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species2unit.R
\name{species2unit}
\alias{species2unit}
\title{Assign standard forest unit based on species composition and additional criteria}
\usage{
species2unit(species_composition, region, ecosite = NA, WG = NA,
  age = NA, site_class = NA, stocking = 1)
}
\arguments{
\item{species_composition}{This can be: (1) A data frame with two columns - first column has to contain species code, the second has to
contain species proportion (species proportions should sum up to 1); (2) a string containing information on species composition.
Note that this string has to be formated as follows: PW 60PR 40, or PW 60 PR 40, or PW60 PR40.}

\item{region}{Region number (1-3)}

\item{ecosite}{Id of ecosite, required for Region 1. Eg.: NW17}

\item{WG}{Working group, required for Region 1.}

\item{age}{Stand age, required for Region 1.}

\item{stocking}{Stocking, required for Region 3. Default = 1}

\item{site_class.}{Site class (1-4), required for Region 2 and 3.}
}
\value{
Standard forest unit code
}
\description{
This function uses a set of rules to assign standard forest unit (SFU or just FU) based on
forest composition, region and additional parameters that depend on the selected region.
}
\examples{

species2unit(species_composition = "PR 90 BW 10", region = 2, site_class = 1)

species2unit(species_composition = "SB  80SW   5LA   5BW   5BF   5", region = 2, site_class = 2)
}
