% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary.comp_fit}
\alias{summary.comp_fit}
\title{Displays a Useful Description of a Fitted Object}
\usage{
\method{summary}{comp_fit}(object, ...)
}
\arguments{
\item{object}{distribution object to summarize.}

\item{...}{additional arguments.}
}
\value{
Function returns summary of the fit, offered by bbmle package for class \code{\link[bbmle]{mle2-class}}.
}
\description{
Displays a useful description of a fitted object.
}
\seealso{
\code{\link[bbmle]{mle2-class}}
}
