% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.R
\name{untrafo}
\alias{untrafo}
\alias{untrafo.trans_standist}
\alias{untrafo.trans_mixdist}
\alias{untrafo.trans_compdist}
\title{Untransformation of a Distribution Object}
\usage{
untrafo(O)

\method{untrafo}{trans_standist}(O)

\method{untrafo}{trans_mixdist}(O)

\method{untrafo}{trans_compdist}(O)
}
\arguments{
\item{O}{transformed distribution object.}
}
\value{
Untransformed distribution object.
}
\description{
\code{untrafo} is a generic function that returns the untransformed random variable, if a transformed object
              is given.
}
\examples{
B  <- binomdist(10, 0.3)
B2 <- -3*log(B)
B2

untrafo(B2)
}
