% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandAttributes.R
\name{StandAttributes}
\alias{StandAttributes}
\title{Calculate stand attributes}
\usage{
StandAttributes(SI, age, species = NA, SFU = NA, stocking = 1,
  Origin = "N", PlantedSpp = NA, bhage = NA)
}
\arguments{
\item{SI}{site index}

\item{age}{stand total age}

\item{species}{Id of dominant species. May be NA (default) if \code{SFU} is provided.}

\item{SFU}{Id of Standard Forest Unit (SFU). May be NA (default) if \code{species} is provided. It is recommended to enter both \code{SFU} and \code{species} if attributes are calculated for SFU.}

\item{stocking}{Stocking. Default=1.}

\item{Origin}{Origin of the stand. Use "N" for natural, "P" for plantation. Default="N".}

\item{PlantedSpp}{Opional information on planted species required for some plantations.}

\item{bhage}{Optional. Breast-height age. If not provided calculated automatically based on species and age.}
}
\value{
A data frame containing the calculated stand attributes.
}
\description{
A wrapper function to calculate all available stand attributes using other \code{MISTR} functions.
}
\details{
This function calculates a set of stand attributes including top height (H), basal area (BA), quadratic mean diameter (QMD),
gross total volume (GTV), gross merchantable volume (GMV), net merchantable volume (NMV), volume-to-basal area ratio (VBAR), and cull.

Attributes may be calculated for:
\itemize{
\item dominant species (argument \code{species} is provided, argument \code{SFU} is NA)
\item standard forest unit (both arguments \code{SFU} and \code{species} are provided (recommended)). If \code{species} is missing, the function will try
to identify the \code{species} automatically, which is possible for some SFUs. If the \code{species} is successfully identified the function
will inform the user and return all stand attributes. If the \code{species} cannot be identified automatically, only a limited set of
attributes will be returned (BA, QMD, N).
}
}
\examples{
#-------------
#Calculate attributes for a 30-year-old White Ash stand, with site index of 20:
StandAttributes(SI = 20, age = 30, species = "AW")

#Calculate attributes for a Red pine stand with site index of 18, for a sequence of ages between 10 and 50:
StandAttributes(SI = 18, age = 10:50, species = "PR")

#Calculate attributes for a stand classified as PJ1 forest unit, dominated by Jack Pine (PJ), with site index of 18, and age of 85
StandAttributes(SI = 18, age = 85, species = "PJ", SFU = "PJ1")

#In some cases the species can be omitted.
StandAttributes(SI = 20, age = 50, SFU = "PJ1")

#In some cases species is required to calculate all attributes. Without species only QMD, BA, and N are calculated.
StandAttributes(SI = 20, age = 50, SFU = "SF1")

#-------------
#Calculating stand attributes for planted stands:
StandAttributes(SI = 20, age = 50, SFU = "PJ1", species = "PJ", Origin = "P")

#For some planted stands PlantedSpp argument is requried
StandAttributes(SI = 20, age = 50, SFU = "SP1", species = "SW", Origin = "P", PlantedSpp = "SW")
}
