% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandAttributes.R
\name{StandAttributes}
\alias{StandAttributes}
\title{Calculate stand attributes}
\usage{
StandAttributes(SI, age, SppId, stocking = 1, Origin = "N",
  PlantedSpp = NA, bhage = age - get_BHA(SppId = SppId))
}
\arguments{
\item{SI}{site index}

\item{age}{stand total age}

\item{SppId}{Id of a species or forest unit}

\item{stocking}{Stocking.}

\item{Origin}{Origin of the stand. Use N for natural, P for plantation.}

\item{PlantedSpp}{Opional information on planted species required for some plantations.}
}
\value{
A data frame containing the calculated stand attributes
}
\description{
A wrapper function to calculate all available stand attributes using other MISTR functions.
}
